backend/requirements.txt
fastapi
uvicorn[standard]
netmiko
xmltodict
pymongo[srv]
python-dotenv
pydantic